requirements.txt
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
flaskwebgui==0.1.9
gunicorn==20.0.4
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
psutil==5.7.0
pylint==2.5.3
requests==2.24.0
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
ufal.udpipe==1.2.0.3
urllib3==1.25.9
Werkzeug==1.0.1
whichcraft==0.6.1
wrapt==1.12.1
